dimens and move packages
