+++ app/src/main/java/com/spbstu/reversemarket/sell/presentation/SellViewModel.kt
@@ -1,4 +1,4 @@
-package com.spbstu.reversemarket.sell
+package com.spbstu.reversemarket.sell.presentation
 
 import androidx.lifecycle.ViewModel
