+++ app/src/main/res/layout/fragment_buy.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".sell.SellFragment">
+    tools:context=".sell.presentation.SellFragment">
 
     <TextView
         android:id="@+id/text_home"
