+++ app/src/main/res/layout/fragment_sell.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F1F1F1"
-    tools:context=".sell.SellFragment">
+    tools:context=".sell.presentation.SellFragment">
 
     <include
         android:id="@+id/frg_search_bar"
@@ -18,6 +18,8 @@
         android:layout_width="match_parent"
         android:layout_height="42dp"
         android:background="#FFFFFF"
+        android:paddingStart="@dimen/big_dimen"
+        android:paddingEnd="@dimen/big_dimen"
         android:layout_below="@id/frg_search_bar"/>
 
     <androidx.recyclerview.widget.RecyclerView
@@ -25,7 +27,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:overScrollMode="never"
-        android:layout_marginBottom="10dp"
+        android:layout_marginBottom="@dimen/big_dimen"
         android:layout_below="@id/frg_tags_list"/>
