+++ app/src/main/res/layout/layout_product_item.xml
@@ -3,9 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginStart="15dp"
-    android:layout_marginTop="15dp"
-    android:layout_marginEnd="15dp"
+    android:layout_marginStart="@dimen/huge_dimen"
+    android:layout_marginTop="@dimen/huge_dimen"
+    android:layout_marginEnd="@dimen/huge_dimen"
     android:background="@color/colorAccent"
     app:cardCornerRadius="5dp">
 
@@ -17,9 +17,8 @@
             android:id="@+id/layout_product_item__name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="15dp"
-            android:layout_marginEnd="20dp"
+            android:layout_marginStart="@dimen/huge_dimen"
+            android:layout_marginTop="@dimen/huge_dimen"
             android:gravity="center_vertical"
             android:lineSpacingExtra="3sp"
             android:text="@string/nike"
@@ -31,8 +30,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@+id/layout_product_item__name"
-            android:layout_marginStart="20dp"
-            android:layout_marginEnd="20dp"
+            android:layout_marginStart="@dimen/huge_dimen"
             android:gravity="bottom"
             android:lineSpacingExtra="3sp"
             android:text="Air Force 1 Shadow White Yellow"
@@ -44,20 +42,10 @@
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__name"
             android:layout_alignParentEnd="true"
-            android:layout_marginEnd="27dp"
+            android:layout_marginEnd="@dimen/def_dimen"
             android:text="150"
-            android:textColor="#828282" />
-
-        <ImageView
-            android:id="@+id/layout_product_item__eye_icon"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/layout_product_item__name"
-            android:layout_alignParentEnd="true"
-            android:layout_marginTop="5dp"
-            android:layout_marginEnd="10dp"
-            android:src="@drawable/ic_eye" />
-
+            android:textColor="#828282"
+            app:drawableEndCompat="@drawable/ic_eye" />
 
         <ImageView
             android:id="@+id/layout_product_item__main_image"
@@ -72,8 +60,8 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__main_image"
-            android:layout_marginStart="10dp"
-            android:layout_marginEnd="10dp" />
+            android:layout_marginStart="@dimen/def_dimen"
+            android:layout_marginEnd="@dimen/def_dimen" />
 
 
         <TextView
@@ -81,8 +69,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__tags"
-            android:layout_marginStart="10sp"
-            android:layout_marginEnd="10sp"
+            android:layout_marginStart="@dimen/def_dimen"
+            android:layout_marginEnd="@dimen/def_dimen"
             android:gravity="top"
             android:lineSpacingExtra="3sp"
             android:text="Nike Air Force 1 - это обновленная версия модели 1982 года со свежими цветовыми решениями и новыми деталями. Этот прочный предмет продолжает..."
@@ -96,8 +84,8 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__description"
-            android:layout_marginStart="10dp"
-            android:layout_marginTop="8dp"
+            android:layout_marginStart="@dimen/def_dimen"
+            android:layout_marginTop="@dimen/def_dimen"
             android:text="@string/price_title"
             android:textColor="#828282"
             android:textSize="14sp"
@@ -108,7 +96,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__price_title"
-            android:layout_marginStart="10sp"
+            android:layout_marginStart="@dimen/def_dimen"
             android:text="$132.10"
             android:textColor="#2E2D2C"
             android:textSize="20sp"
@@ -120,8 +108,8 @@
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__description"
             android:layout_alignParentEnd="true"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="10sp"
+            android:layout_marginTop="@dimen/def_dimen"
+            android:layout_marginEnd="@dimen/def_dimen"
             android:text="UserName"
             android:textSize="13sp"
             android:textStyle="normal" />
@@ -132,8 +120,8 @@
             android:layout_height="wrap_content"
             android:layout_below="@id/layout_product_item__username"
             android:layout_alignParentEnd="true"
-            android:layout_marginTop="6dp"
-            android:layout_marginEnd="10sp"
+            android:layout_marginTop="@dimen/small_dimen"
+            android:layout_marginEnd="@dimen/def_dimen"
             android:text="20.09.2020"
             android:textSize="13sp"
             android:textStyle="normal" />
