+++ app/src/main/res/layout/layout_product_tag.xml
@@ -7,11 +7,11 @@
         android:id="@+id/layout_product_tag__name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:paddingTop="6dp"
-        android:paddingBottom="6dp"
-        android:layout_marginTop="2dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginEnd="10dp"
+        android:paddingTop="@dimen/small_dimen"
+        android:paddingBottom="@dimen/def_dimen"
+        android:layout_marginTop="@dimen/small_dimen"
+        android:layout_marginStart="@dimen/def_dimen"
+        android:layout_marginEnd="@dimen/def_dimen"
         android:text="Кроссовки"
         android:textColor="#2E2D2C"
         android:textSize="16sp"
