+++ app/src/main/res/layout/layout_removable_product_tag.xml
@@ -2,15 +2,13 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:background="@drawable/tag"
-    android:layout_marginStart="12dp"
-    >
+    android:background="@drawable/tag">
         <TextView
             android:id="@+id/layout_removable_product_tag__name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:paddingTop="6dp"
-            android:layout_marginStart="10sp"
+            android:paddingTop="@dimen/def_dimen"
+            android:layout_marginStart="@dimen/def_dimen"
             android:text="Кроссовки"
             android:textColor="#2E2D2C"
             android:textSize="16sp"
@@ -22,9 +20,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:src="@drawable/ic_delete_tag_image"
-            android:padding="9dp"
-            android:paddingStart="6dp"
-            android:paddingEnd="6dp"
+            android:padding="@dimen/def_dimen"
             />
 
 </LinearLayout>
