+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,10 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+
+    <dimen name="small_dimen">4dp</dimen>
+    <dimen name="def_dimen">8dp</dimen>
+    <dimen name="big_dimen">12dp</dimen>
+    <dimen name="huge_dimen">16dp</dimen>
+
 </resources>
